[style] Sort stops dictionaries before conversion to expression (#81)
* [style] Sort stops dictionaries before conversion to expression

* Handle negative values in dictionary

+++ Examples/Examples/All Examples/ColorExpressionExample.swift
@@ -44,7 +44,7 @@ public class ColorExpressionExample: UIViewController, ExampleProtocol {
          ]
          */
 
-        let stops: [UInt: UIColor] = [
+        let stops: [Double: UIColor] = [
             0: .red,
             14: .blue
         ]
